naver-content-analytics/requirements.txt
pandas==2.2.3
SQLAlchemy==2.0.38
psycopg2-binary==2.9.10
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
